change traffic->max-burst config
Increase max-burst to support
flows with high rate.

+++ code/bngblaster/src/bbl_ctx.h
@@ -334,7 +334,7 @@ typedef struct bbl_ctx_
         bool stream_rate_calc; /* Enable/disable stream rate calculation */
         bool stream_delay_calc; /* Enable/disable stream delay calculation */
         bool stream_udp_checksum; /* Enable/disable stream UDP checksum calculation */
-        uint8_t stream_max_burst; /* Limit the max packets per TX interval */
+        uint16_t stream_max_burst; /* Limit the max packets per TX interval */
 
         /* Session Traffic */
         bool session_traffic_autostart;
